wish generator refactor
