+++ lib/screens/birthday_info_page.dart
@@ -1,10 +1,9 @@
-import 'dart:math';
 import 'package:daty/components/precise_age.dart';
 import 'package:daty/components/view_title.dart';
-import 'package:flutter/services.dart';
 
 import 'package:confetti/confetti.dart';
 import 'package:daty/components/birthday_timer/birthday_timer.dart';
+import 'package:daty/components/wish_generator.dart';
 import 'package:daty/screens/birthday_edit_page.dart';
 import 'package:daty/utilities/Birthday.dart';
 import 'package:daty/utilities/birthday_data.dart';
@@ -62,7 +61,7 @@ class _BirthdayInfoPageState extends State<BirthdayInfoPage> {
               BirthdayTimer(widget.birthdayId),
               const SizedBox(height: 40),
               ViewTitle(AppLocalizations.of(context)!.generateWish),
-              wishDisplay(),
+              WishGenerator(getDataById(widget.birthdayId)),
               const SizedBox(height: 10),
               allowNotificationSwitch(),
             ],
@@ -242,83 +241,6 @@ class _BirthdayInfoPageState extends State<BirthdayInfoPage> {
         ),
       ),
     );
-  }
-
-  Widget wishDisplay() {
-    String birthdayWish = getWish();
-
-    return Padding(
-      padding: const EdgeInsets.only(right: 40, left: 40, top: 20),
-      child: Container(
-        decoration: const BoxDecoration(
-          color: Constants.darkGreySecondary,
-          borderRadius: BorderRadius.all(Radius.circular(10)),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Row(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              Expanded(
-                child: Padding(
-                  padding: const EdgeInsets.all(20),
-                  child: Text(
-                    birthdayWish,
-                    textAlign: TextAlign.center,
-                    style: const TextStyle(
-                      color: Constants.whiteSecondary,
-                      fontSize: Constants.normalFontSize,
-                    ),
-                  ),
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.only(right: 20),
-                child: Column(
-                  children: [
-                    ElevatedButton(
-                      onPressed: () async {
-                        await Clipboard.setData(
-                          ClipboardData(text: birthdayWish),
-                        );
-                      },
-                      child: const Icon(Icons.copy),
-                    ),
-                    ElevatedButton(
-                      onPressed: (() {
-                        setState(() {});
-                      }),
-                      child: const Icon(Icons.refresh),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  String getWish() {
-    List<String> wishes = [
-      AppLocalizations.of(context)!.wish1,
-      AppLocalizations.of(context)!.wish2,
-      AppLocalizations.of(context)!.wish3,
-      AppLocalizations.of(context)!.wish4,
-      AppLocalizations.of(context)!.wish5,
-    ];
-
-    int number = Random().nextInt(wishes.length);
-    String wish = wishes[number];
-
-    String name = getDataById(widget.birthdayId).name;
-    wish = wish.replaceAll('/name/', name);
-
-    int age = Calculator.calculateAge(getDataById(widget.birthdayId).date) + 1;
-    wish = wish.replaceAll('/age/', age.toString());
-
-    return wish;
   }
 
   Padding allowNotificationSwitch() {
